requirements.txt
Flask==2.3.3
nibabel==5.1.0
numpy==1.24.3
Pillow==10.0.0
matplotlib==3.7.2
torch>=1.9.0
torchvision>=0.10.0
scikit-image>=0.18.0
scipy>=1.6.0
tqdm>=4.60.0